Refactor of TestVspherePoolDriver
- Split up into TestVspherePoolDriverConfiguration and TestVspherePoolDriverOperations
- Using TestUtils instead of private method for reading sample configs
