+++ vsphere/src/test/java/com/elastisys/scale/cloudpool/vsphere/driver/TestVspherePoolDriverOperations.java
@@ -1,0 +1,53 @@
+package com.elastisys.scale.cloudpool.vsphere.driver;
+
+import static org.junit.Assert.fail;
+
+import org.junit.Test;
+
+public class TestVspherePoolDriverOperations {
+
+    @Test
+    public void testVspherePoolDriver() {
+        fail("Not yet implemented");
+    }
+
+    @Test
+    public void testListMachines() {
+        fail("Not yet implemented");
+    }
+
+    @Test
+    public void testStartMachines() {
+        fail("Not yet implemented");
+    }
+
+    @Test
+    public void testTerminateMachine() {
+        fail("Not yet implemented");
+    }
+
+    @Test
+    public void testAttachMachine() {
+        fail("Not yet implemented");
+    }
+
+    @Test
+    public void testDetachMachine() {
+        fail("Not yet implemented");
+    }
+
+    @Test
+    public void testSetServiceState() {
+        fail("Not yet implemented");
+    }
+
+    @Test
+    public void testSetMembershipStatus() {
+        fail("Not yet implemented");
+    }
+
+    @Test
+    public void testGetPoolName() {
+        fail("Not yet implemented");
+    }
+}
